In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,14 +9,14 @@
 int main(void)
 {
 	int hub = 0;
-	unsigned long int jub = 0, kub = 1, next = 0;
+	unsigned long long int jub = 0, kub = 1, next = 0;
 
 	while (hub < 98)
 	{
 		next = jub + kub;
 		jub = kub;
 		kub = next;
-		printf("%lu", next);
+		printf("%llu", next);
 
 		if (hub < 97)
 		{
